paint over original image
